Implementazione moltiplicatori
Aggiunto funzionamento e upgrade per i moltiplicatori di danno e moneta.

+++ Assets/Scripts/Shop/Coin.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public class Coin : MonoBehaviour
 {
-    public int value = 5;
+    public float value = 5;
 }
